DRY: Paginated mixin to make page getter/setter reusable
